Multilinear

Description

Defines main tyclasses to deal with multilinear algebra and re-exports content of other library modules.

Classes

class Multilinear t where

MULTILINEAR CLASS - TENSOR DEFINED AS MULTILINEAR MAP

Minimal complete definition

(!+), (!-), (!*)

Methods

(!+) :: t -> t -> t infixl 7

Tensor sum

(!-) :: t -> t -> t infixl 7

Tensor difference

(!*) :: t -> t -> t infixl 8

Tensor product satisfying Einstein summation convention

class Num t => Tensor t where

TENSOR CLASS - TENSOR AS MULTIDIMENSIONAL ARRAY

Minimal complete definition

generate, fromList, order, elems, indices, rename, equiv, transpose, transpose1, concat

Methods

generate :: TIndex -> (Int -> t) -> t

Generate tensor elements as a function of indices

fromList :: TIndex -> [a] -> t

generate rank tensor from a list

order :: t -> (Int, Int)

Tensor order (contravariant, covariant)

elems :: t -> Int

Number of tensor elements

indices :: t -> [TIndex]

List of tensor indices

rename :: t -> String -> String -> t

Rename tensor index

equiv :: t -> t -> Bool

Check if tensors are equivalent (are of the same type and size)

transpose :: t -> t

switch all indices type

transpose1 :: t -> t

switch only one index type

concat :: TIndex -> t -> t -> t

concatenation of two tensor with given index or by creating a new one

Other library modules