Multilinear.Generic

Description

This module contains generic implementation of tensor defined as nested arrays

Documentation

data Tensor a

Tensor defined recursively as scalar or array of other tensors

c is type of a container, i is type of index size and a is type of tensor elements

Constructors

Scalar
Scalar

Fields

scalarVal :: a

value of scalar

SimpleFinite
Simple, one-dimensional finite tensor

Fields

tensorFiniteIndex :: Index

Finite index Mutltilinear.Index.Finite of tensor

tensorScalars :: Vector a

FiniteTensor
Finite array of other tensors

Fields

tensorFiniteIndex :: Index

Finite index Mutltilinear.Index.Finite of tensor

tensorsFinite :: Vector (Tensor a)

Array of tensors on deeper recursion level

InfiniteTensor
Infinite list of other tensors

Fields

tensorInfiniteIndex :: Index

Infinite index Mutltilinear.Index.Infinite of tensor

tensorsInfinite :: [Tensor a]

Infinite list of tensors on deeper recursion level

Err
Operations on tensors may throw an error

Fields

errMessage :: String

Error message

(!)

Arguments

:: Tensor a
tensor t

-> Int
index i

-> Tensor a
tensor t[i]

Recursive indexing on list tensor t ! i = t[i]

mergeScalars :: Tensor a -> Tensor a

Merge FiniteTensor of Scalars to SimpleFinite tensor for performance improvement

isScalar :: Tensor a -> Bool

Return true if tensor is a scalar

isSimple :: Tensor a -> Bool

Return true if tensor is a simple tensor

isFiniteTensor :: Tensor a -> Bool

Return True if tensor is a complex tensor

isInfiniteTensor :: Tensor a -> Bool

Return True if tensor is a infinite tensor

dot

Arguments

:: (Num a, NFData a)

=> Tensor a
First dot product argument

-> Tensor a
Second dot product argument

-> Tensor a
Resulting dot product

_elemByElem

Arguments

:: (Num a, NFData a)

=> Tensor a
First argument of operator

-> Tensor a
Second argument of operator

-> (a -> a -> a)
Operator on tensor elements if indices are different

-> (Tensor a -> Tensor a -> Tensor a)
Tensor operator called if indices are the same

-> Tensor a
Result tensor

Apply a tensor operator elem by elem. For time & memory optimization, first push common indices right and mergeScalars of the result

contractionErr

Arguments

:: TIndex
Index of first dot product parameter

-> TIndex
Index of second dot product parameter

-> Tensor a
Erorr message

tensorIndex :: Tensor a -> TIndex

Return generic tensor index