Multilinear.Index

Description

Generic tensor index which may be finitely- or infinitely-dimensional.

Documentation

class Index i where

Tensor index class which may be lower (covariant), upper (contravariant) or indifferent.

Minimal complete definition

indexName, isCovariant, isContravariant, isIndifferent, equivI, toTIndex

Methods

indexName :: i -> String

Index name

isCovariant :: i -> Bool

Returns True if index is lower (covariant), False otherwise.

isContravariant :: i -> Bool

Returns True if index is upper (contravariant), False otherwise.

isIndifferent :: i -> Bool

Returns True if index if indifferent, False otherwise.

equivI :: i -> i -> Bool

Returns True if two indices are equivalent, thus differs only by name, but share same size and type.

(!=!) :: i -> i -> Bool infixl 2

Infix equivalent for equiv. Has low priority equal to 2.

toTIndex :: i -> TIndex

Convert to generic index type

data TIndex

Generic index type finitely- or infinitely-dimensional

Constructors

Covariant

Fields

indexSize :: Maybe Int

tIndexName :: String

Contravariant

Fields

indexSize :: Maybe Int

tIndexName :: String

Indifferent

Fields

indexSize :: Maybe Int

tIndexName :: String