Multilinear.Vector

Description

This module provides convenient constructors that generates a vector (tensor with one upper index) of finite or infinite size.

Finitely-dimensional vectors provide much greater performance than infinitely-dimensional

Generators

fromIndices

Arguments

:: Num a

=> String
Index name (one character)

-> Int
Number of elements

-> (Int -> a)
Generator function - returns a vector component at index i

-> Tensor a
Generated vector

Generate vector as function of indices

const

Arguments

:: Num a

=> String
Index name (one character)

-> Int
Number of elements

-> a
Value of each element

-> Tensor a
Generated vector

Generate vector with all components equal to some v

randomDouble

Arguments

:: ContGen d

=> String
Index name (one character)

-> Int
Number of elements

-> d
Continuous probability distribution (as from Statistics.Distribution)

-> IO (Tensor Double)
Generated vector

Generate vector with random real components with given probability distribution. The vector is wrapped in the IO monad.

Available probability distributions:

Beta : Statistics.Distribution.BetaDistribution

Cauchy : Statistics.Distribution.CauchyLorentz

Chi-squared : Statistics.Distribution.ChiSquared

Exponential : Statistics.Distribution.Exponential

Gamma : Statistics.Distribution.Gamma

Normal : Statistics.Distribution.Normal

StudentT : Statistics.Distribution.StudentT

Uniform : Statistics.Distribution.Uniform

F : Statistics.Distribution.FDistribution

Laplace : Statistics.Distribution.Laplace

randomDoubleSeed

Arguments

:: (ContGen d, PrimMonad m)

=> String
Index name (one character)

-> Int
Number of elements

-> d
Continuous probability distribution (as from Statistics.Distribution)

-> Int
Randomness seed

-> m (Tensor Double)
Generated vector

Generate vector with random real components with given probability distribution and given seed. The vector is wrapped in a monad.

Available probability distributions:

Beta : Statistics.Distribution.BetaDistribution

Cauchy : Statistics.Distribution.CauchyLorentz

Chi-squared : Statistics.Distribution.ChiSquared

Exponential : Statistics.Distribution.Exponential

Gamma : Statistics.Distribution.Gamma

Normal : Statistics.Distribution.Normal

StudentT : Statistics.Distribution.StudentT

Uniform : Statistics.Distribution.Uniform

F : Statistics.Distribution.FDistribution

Laplace : Statistics.Distribution.Laplace

randomInt

Arguments

:: DiscreteGen d

=> String
Index name (one character)

-> Int
Number of elements

-> d
Discrete probability distribution (as from Statistics.Distribution)

-> IO (Tensor Int)
Generated vector

Generate vector with random integer components with given probability distribution. The vector is wrapped in the IO monad.

Available probability distributions:

Binomial : Statistics.Distribution.Binomial

Poisson : Statistics.Distribution.Poisson

Geometric : Statistics.Distribution.Geometric

Hypergeometric: Statistics.Distribution.Hypergeometric

randomIntSeed

Arguments

:: (DiscreteGen d, PrimMonad m)

=> String
Index name (one character)

-> Int
Number of elements

-> d
Discrete probability distribution (as from Statistics.Distribution)

-> Int
Randomness seed

-> m (Tensor Int)
Generated vector

Generate vector with random integer components with given probability distribution and given seed. The vector is wrapped in a monad.

Available probability distributions:

Binomial : Statistics.Distribution.Binomial

Poisson : Statistics.Distribution.Poisson

Geometric : Statistics.Distribution.Geometric

Hypergeometric: Statistics.Distribution.Hypergeometric

From files

fromCSV

Arguments

:: (Num a, NFData a, Serialize a)

=> String
Index name (one character)

-> String
CSV file name

-> Char
Separator expected to be used in this CSV file

-> EitherT SomeException IO (Tensor a)
Generated vector or error message

Read vector components from CSV file. Reads only the first row of the file.

toCSV

Arguments

:: (Num a, NFData a, Serialize a)

=> Tensor a
vector to serialize

-> String
CSV file name

-> Char
Separator expected to be used in this CSV file

-> IO Int
Number of rows written

Write vector to CSV file.